+++ integrations/snyk/CHANGELOG.md
@@ -7,11 +7,26 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 <!-- towncrier release notes start -->
 
+# Port_Ocean 0.1.38 (2024-04-04)
+
+### Bug Fixes
+
+- Fixed request json error in snyk client and removed code section that handles SNYK-9999 code.
+
+
+# Port_Ocean 0.1.37 (2024-04-03)
+
+### Bug Fixes
+
+- Fixed a bug in the pagination logic that caused the integration to return SNYK-9999 error
+
+
 # Port_Ocean 0.1.36 (2024-04-03)
 
 ### Bug Fixes
 
 - Fixed handling of code SNYK-9999 from Snyk Api
+
 
 # Port_Ocean 0.1.35 (2024-04-02)